requirements.txt
unstructured[local-inference]
py-llm-core
